正在写批量删除菜品接口

+++ sky-server/src/main/java/com/sky/mapper/DishMapper.java
@@ -20,5 +20,5 @@ public interface DishMapper {
 
     Page<DishVO> page(Dish employee);
 
-    void deleteBatch(List<Integer> dishIds);
+    void deleteBatch(List<Long> dishIds);
 }
